Fix a bug that has been around for a while.
The iospace region was not getting created like it should. Back when I introduced the overcommit code, the large region was not getting allocated. Problem is it's size was being committed from the page pool. No good, since no pages will ever be attached to the cache associated with that region. It's really just there for a placeholder, so that section of the kernel doesn't get fooled with.

Solution: Create a new null vm_store object that just returns a fatal error if a page fault hits it. That's what we want, since the iospace region should never get a page fault. This is a better solution anyway, since the old system would have ended up with a few pages attached to it if a fault happened, which is no good.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@314 c25cc9d1-44fa-0310-b259-ad778cb1d433