integrate the latest copy of the hoard allocator (found at http://www.hoard.org). Now user space can malloc and all that.
The integration has a pretty high hack factor, due to the lack of a real sbrk. I'll probably redo it to either extend the current region, or create a bunch of other regions. Time will tell.

Also, this is the first time any c++ has entered the system. crt0 had to be modified to deal with ctors, and some of the librarys were wrapped with the proper extern "C" wrapper, so c++ can probably be in general used in user space, except for exceptions, which I may enable soon.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@321 c25cc9d1-44fa-0310-b259-ad778cb1d433