Added the beginning of a beos driver layer. It'll be largely a add-as-you-need kind of operation, but as of now at least one beos driver loads and runs: speaker
It's not included in the build, and it wont be, but if it's placed in addons/beosdev/speaker on the bootfs volume, it'll be picked up and loaded. This is accomplished by editing a line in boot/i386/config.ini and adding speaker.

Just a start, but I should try for some more complex drivers.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@333 c25cc9d1-44fa-0310-b259-ad778cb1d433