Bring the Dreamcast port back up to build/run status
Doesn't completely run with this checkin, there's some subtle bug that causes it to bail when it tries to start init. Lots of the sh4 code experienced some code rot, so I need to walk back through it and clean it up. I'll put that off for now, but it makes it pretty far into the kernel. Also, I have a problem with newer gcc 3 that causes it to not produce a kernel with a dynamic section. I need to figure that out, but I'll do it on the ia-32 compile and roll the changes back to here.

Many of these files were simply touched while making it work, and are worth checking in to clean up the formatting. I have my editor in a mode that removes trailing spaces on lines, which is a worthy cause.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@338 c25cc9d1-44fa-0310-b259-ad778cb1d433