added cbuf code to allocate from it's own pool of memory. Currently doesn't free it when it's done, but that requires a hook into the vm that does not exist yet.
cbufs are essentially just like mbufs in bsd. At present, they will be used in the upcoming network stack and the beos-style ports system currently being developed. They have the advantage of allowing the cbuf pool to become fragmented, in addition to the standard mbuf advantages (adding to the front of buffer, trimming from either end of the buffer, etc).


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@358 c25cc9d1-44fa-0310-b259-ad778cb1d433