three patches by Jeroen Van Iddekinge:
1) Redirect bug. 
File  : apps/shell/command.c (old version shell)

When output of command is redirected to the a file and the file didn't exists the command failed. With this patch, the file is now created when it doesn't exists.

2) printf bug
file  : lib/libsys/stdio.c

The 'file position' parameter of sys_write in the printf function 
is changed from 0 to -1. When this parameter was 0 all lines outputed 
are placed at file position 0.

3) Buffer overflow 
file  : kernel/vfs.c

In path_to_vnode_dir  ".\0" is placed after the last '\' in
 the path name.  This can  in sometimes  cause a buffer overflow when 
this path name ends with a "\"


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@476 c25cc9d1-44fa-0310-b259-ad778cb1d433