total hack to make gcc 3.x happy.
when compiling the hoard code which uses c++, it always emits a reference to a function that's supposed to be in libstdc++, which we dont have here. 
I've stuck in a null version of that function, with the idea that it'll never get called, since it's an exception utility function.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@495 c25cc9d1-44fa-0310-b259-ad778cb1d433