Checking in another version of the IDE driver by Rob Judd.
This is a modification of the one that's in the tree, which has a compilicated lineage. In order to deal with the parallel development of the driver, I'm just going to check in multiple copies of it, and the best one will win.

Here are some notes Rob sent with it:

It will recognise 6 of 8 partitions on my main drive (the other two being NTFS) and the first CD-ROM drive on the second controller. These are all registered by devfs and available to drivers. A bit more tidying of the code was also accomplished.

Subtle changes:

1) hdd partitions are registered as /dev/disk/ata/x/x/x (omitting the 
redundant /dev/disk/ide/ path)
2) cd drives are registered as /dev/disk/atapi/x/x
3) all known extended partition types (0x05, 0x0f, 0x85) are catered 
for in the code, and recognition of Ext DOS partitions is tested as 
working

Bugs:

1) Second drives on the same controller can't be seen.
2) Whether the code works or not depends on whether certain dprintf() 
statements are present. I suspect this is a memory issue, but haven't 
been able to nail it down.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@580 c25cc9d1-44fa-0310-b259-ad778cb1d433