Here we have it folks! Step right up, new make system!
Replaced a complicated make system with something that is probably just as complicated,
but a little more maintainable. It has a few side effects:

1) All target files are built into a seperate directory structure (build/i386/..., build/sh4/...). 
The obj.* directories are depreciated. Delete them, or it'll confuse the hell out of you. The final
image is at build/<ARCH>/final, where ARCH is probably 'i386'.

2) 'make' now builds a floppy image, instead of having to do 'make floppy'. The reason it used to be
that way was lost in obscurity.

3) Probably some other stuff. Lots of folks will probably be screwed up with this.

Hopefully there shouldn't be too many bugs, but if you see some weird stuff, dont hesitate to bug me
about it. I've tested it on all my machines with gmake 3.79.1, so earlier versions may not work.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@607 c25cc9d1-44fa-0310-b259-ad778cb1d433