mutex change:
The mutex lock that I've been using was a lie. It didn't act that much like a real mutex, it was actually a benephore, which is just a semaphore protected by an atomic var. Doesn't make too much sense in kernel space.

So, to make mutexes be more like mutexes, it now stores the holder thread, and has a couple of restrictions:
1) the thread releasing a mutex *must* be the thread that acquired it. mutexes are totally designed for critical section stuff, not general use. 
2) the mutex will detect being acquired twice by itself, which is nice for debugging.
3) since the holder thread field is initialized to -1, there is apretty good chance that if you use a mutex structure without initializing it, it'll die, and I've caught some bugs with this already.

As a result of implementing this I found at least one serious bug. This one was in the vm.c code that reinitialized all the vm structures' semaphores after the sem code came up. It was setting up the mutexes with the wrong count. May be the cause of some of the instability under heavy load.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@665 c25cc9d1-44fa-0310-b259-ad778cb1d433