tcp/cbuf optimization:
build the write buffer not by adding a new cbuf to the end of the write buffer chain on
every tcp write. Extend the old chain and add to it. Make a *massive* difference in memory 
savings. Moved the cbuf size to 2048 again, since we're not wasting much memory any more.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@816 c25cc9d1-44fa-0310-b259-ad778cb1d433