Munge the libc header files around enough so that gcc can build libstdc++ against it. A pretty major milestone actually.
Next step is to remove the local libsupc++ library and c++ header files,
and figure out precisely how to build gcc to have this all precompiled.

Also, the headers are laid out pretty close to how they're supposed to,
might be easy enough to port some apps directly to newos.


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@1127 c25cc9d1-44fa-0310-b259-ad778cb1d433