Pressing a key before the PS/2 mouse driver was loaded resulted in system hang - very unfortunate.
The reason is that on pressing a key the output buffer was filled with its code, but the initialization required the output buffer to be empty.
This is resolved now by reading from the output buffer before initialization.
Note that there is a small race condition, but for now it should be ok.
Does writing a control byte really require the output buffer to be empty?


git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@1155 c25cc9d1-44fa-0310-b259-ad778cb1d433