Get more of the dreamcast port working.
Moved the kernel into the P1 area (physical memory is mapped at 0x80000000). This
greatly simplifies the stage2 loader. Turned off the vcpu stuff, which means we'll
have to hack real mmu support in the kernel, and some mechanism to let the VM deal with
areas of kernel space that cannot be controlled by the mmu (0xc0000000 - 0xdfffffff is the only
priviledged area that can be mmued on an SH4).



git-svn-id: svn+ssh://newos.org/var/svn/newos/newos@1205 c25cc9d1-44fa-0310-b259-ad778cb1d433